requirements.txt
markdown
beautifulsoup4
langchain
langchain_community
faiss-cpu
transformers
fastapi
uvicorn
pydantic
pyyaml
orjson
torch
numpy
requests
watchdog
textual
rank_bm25
google-genai